style: Change Navbar styles
Navbar Desktop & Mobile/tablet style changed

PEIT-25 Ajustar Navbar
